Split AutoBank feature into it's own plugin.
Remove logic handling autobanking from MQ2AutoInventory.cpp and renaming a couple of functions to allow remaining features to continue to function without conflicting with the split out plugin.
