+++ src/MacroQuest.sln
@@ -58,6 +58,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{EAFB7791
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NamedPipeClient", "tests\NamedPipeClient\NamedPipeClient.vcxproj", "{312C5DE6-34C8-4474-B186-12989694C780}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MQ2AutoBank", "plugins\autobank\MQ2AutoBank.vcxproj", "{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Win32 = Debug|Win32
@@ -222,6 +224,14 @@ Global
 		{312C5DE6-34C8-4474-B186-12989694C780}.Debug|x64.ActiveCfg = Debug|x64
 		{312C5DE6-34C8-4474-B186-12989694C780}.Release|Win32.ActiveCfg = Release|Win32
 		{312C5DE6-34C8-4474-B186-12989694C780}.Release|x64.ActiveCfg = Release|x64
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Debug|Win32.ActiveCfg = Debug|Win32
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Debug|Win32.Build.0 = Debug|Win32
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Debug|x64.ActiveCfg = Debug|x64
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Debug|x64.Build.0 = Debug|x64
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Release|Win32.ActiveCfg = Release|Win32
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Release|Win32.Build.0 = Release|Win32
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Release|x64.ActiveCfg = Release|x64
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -251,6 +261,7 @@ Global
 		{72EE75F4-BCFA-4152-BFC6-A3C2A2B2C9AC} = {42D9994B-93C6-4C4B-971A-A7C918CA4DB8}
 		{EAFB7791-F141-4B87-A0F9-B5685A90A2C1} = {42D9994B-93C6-4C4B-971A-A7C918CA4DB8}
 		{312C5DE6-34C8-4474-B186-12989694C780} = {EAFB7791-F141-4B87-A0F9-B5685A90A2C1}
+		{C0E145AB-4882-4FD4-8ADD-630FC678FBC0} = {A648B03F-7642-4857-A62A-AFABC7CAB451}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {330AC4A2-17BC-4784-AB3C-2B1DA71EB6A5}
