+++ src/main/MQ2PluginHandler.cpp
@@ -618,7 +618,7 @@ void PluginsCleanUI()
 	PluginDebug("PluginsCleanUI()");
 
 	DeleteMQ2NewsWindow();
-	RemoveAutoBankMenu();
+	RemoveFindItemMenu();
 
 	ForEachPlugin([](const MQPlugin* plugin)
 		{
