+++ src/plugins/autobank/README.md
@@ -1,0 +1,44 @@
+# MQ2AutoBank
+AutoBank adds a context menu to the AutoBank button in EQ's bank window. It allows you to automatically remove or add items to the bank with Tradeskill, Collectible, or Quest tags on it. 
+Additionally, in the case of tradeskill items you can filter if you should or shouldn't move items with skill mods (Trophies). 
+More feature to be considered/added in the future. 
+
+## Getting Started
+
+Quick start instructions to get users up and going
+
+```txt
+/plugin MQ2AutoBank
+```
+
+### Commands
+```txt
+None
+```
+
+### Configuration File
+
+Describe the configuration file and what the settings do
+
+```Macroquest.ini
+[AutoBank]
+AutoBankCollectibleItems=0
+AutoBankTradeSkillItems=0
+AutoInventoryItems=0
+AutoBankTropiesWithTradeskill=0
+AutoBankTrophiesWithTradeskill=1
+```
+
+## Authors
+
+* **Eqmule** - *Initial work*
+* **ChatWithThisName**
+* **Brainiac**
+
+
+See also the list of [contributors](https://github.com/your/project/contributors) who participated in this project.
+
+## Acknowledgments
+
+* Inspiration from...
+* I'd like to thank the Thieves' Guild for helping me with all the code I stole...
